requirement.txt
django
djangorestframework
psycopg2-binary
drf-yasg
python-decouple
django-filter
django-ckeditor
djangorestframework-simplejwt
pillow
django-cors-headers
